Move test_helpers.hpp to public interface
