+++ rviz_default_plugins/test/rviz_default_plugins/page_objects/path_display_page_object.cpp
@@ -34,7 +34,7 @@
 
 #include <QTest>  // NOLINT
 
-#include "rviz_visual_testing_framework/internal/test_helpers.hpp"
+#include "rviz_visual_testing_framework/test_helpers.hpp"
 
 PathDisplayPageObject::PathDisplayPageObject(
   int display_id, std::shared_ptr<Executor> executor,
