+++ rviz_visual_testing_framework/src/page_objects/base_page_object.cpp
@@ -35,7 +35,7 @@
 
 #include <QTest>  // NOLINT
 
-#include "rviz_visual_testing_framework/internal/test_helpers.hpp"
+#include "rviz_visual_testing_framework/test_helpers.hpp"
 
 BasePageObject::BasePageObject(
   int display_id,
